+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## v1.1.38 (1.20.1)
 - added faster mining modifier
 - adjusted default enchanting glint color
+- minor changes to disable glint shader registration as well if glint is disabled in the config